notification_server/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.3
charset-normalizer==3.2.0
click==8.1.7
discord==2.3.2
discord.py==2.3.2
Flask==3.0.0
frozenlist==1.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
priority==2.0.0
python-dotenv==1.0.0
Quart==0.19.3
tomli==2.0.1
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.2
